Removed useless method in Room and update price in RoomType

+++ src/main/java/model/room/RoomType.java
@@ -39,7 +39,7 @@ public enum RoomType {
         return this.price;
     }
 
-    public double getBeds() {
+    public int getBeds() {
         return this.beds;
     }
